requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
cmake==3.30.2
facenet-pytorch==2.3.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
huggingface-hub==0.24.5
idna==3.7
jinja2==3.1.4
joblib==1.4.2
lightning==2.3.3
lightning-utilities==0.11.6
lit==18.1.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==24.1
pillow==10.4.0
pygments==2.18.0
pytorch-lightning==2.3.3
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.4
scikit-learn==1.3.2
scipy==1.10.1
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.0.0
torchaudio==2.0.1
torchmetrics==1.4.1
torchvision==0.15.1
tqdm==4.66.5
transformers==4.40.0
triton==2.0.0
typing-extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
